splitting singleCard again
